Rename mopidy.mixers to mopidy.audio.mixers
